removed before and before.styled to ActivityDetail
