+++ src/components/styled/act-detail-style-comp/Before.styled.jsx
@@ -1,14 +1,14 @@
-import styled from "styled-components";
+// import styled from "styled-components";
 
-export const BeforeContainer = styled.div`
-  margin-bottom: 20px;
-  a {
-    width: 227px;
-    height: 35px;
-    font-family: Poppins;
-    font-size: 24px;
-    text-decoration: none;
-    color: black;
-    font-weight: bold;
-  }
-`;
+// export const BeforeContainer = styled.div`
+//   margin-bottom: 20px;
+//   a {
+//     width: 227px;
+//     height: 35px;
+//     font-family: Poppins;
+//     font-size: 24px;
+//     text-decoration: none;
+//     color: black;
+//     font-weight: bold;
+//   }
+// `;
