+++ src/components/styled/act-detail-style-comp/Common.jsx
@@ -1,5 +1,18 @@
 import styled from "styled-components";
 
+export const BeforeContainer = styled.div`
+  margin-bottom: 20px;
+  a {
+    width: 227px;
+    height: 35px;
+    font-family: Poppins;
+    font-size: 24px;
+    text-decoration: none;
+    color: black;
+    font-weight: bold;
+  }
+`;
+
 export const JoinButton = styled.button`
   width: 242px;
   height: 45px;
